Prototyped capturing the map as an image and drawing it on a SceneKit plane, but this isn't entirely workable.
I had the idea to tap a seismic event map pin, animate to it, and then look beneath the map to see a diagram of the epicenter depth and spheres of wave propagation. But it didn't work out. MapKit and SceneKit's out-of-process drawing conspired to make this too hard to do properly.

- Although the map tells me it is done moving into position and done rendering, it's not done rendering *to the screen*. When I take the snapshot image on my iPhone 6 it is still fading between the original blurry map text and the new, zoomed-in sharp map text. I would have to delay by some arbitrary amount of time before taking the image.
- When the scene view is shown, there is a flash before the map plane with the snapshot image shows up, during which nothing at all is rendering. I'd have to figure out how to wait until SceneKit has drawn its first frame, and I have never even tried SceneKit.